requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.3
django-bootstrap-ui==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-bulk==0.2.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pyserial==3.4
pytz==2019.3
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
zope.interface==4.7.1
